estás ai estás

+++ Aula1/ex3/ex3.c
@@ -23,12 +23,15 @@ int _tmain(int argc, LPTSTR argv[]) {
 		_tprintf(result);
 		fflush(stdin);
 		_fgetts(str, MAX, stdin);
+
 		//Retirar \n
 		str[_tcslen(str) - 1] = '\0';
+		
 		//Maiúsculas
 		for (i = 0; i < _tcslen(str); i++)
 			str[i] = _totupper(str[i]);
-		_stprintf_s(result, MAX, TEXT("Frase:%s, Tamanho:%d\n"), str, _tcslen(str));
+		_stprintf_s(result, MAX, TEXT("Frase:%s, Tamanho:%d\n"), str, (int) _tcslen(str));
 	} while (_tcsicmp(TEXT("FIM"), str));
+
 	return 0;
 }
